zaciatok prace na klientovi, MySocketClient.h
