+++ CMakeLists.txt
@@ -17,13 +17,15 @@ set(COMMON_SOURCES
         utils/SimulationCsvLoader.h
         models/SimulationCsvRecord.h
         utils/MySocketServer.h
-        )
+        utils/MySocketClient.h
+)
 
 # Client executable
 add_executable(pos_sp_client
         ${COMMON_SOURCES}
         main_client.cpp
-        )
+        utils/MySocketClient.h
+)
 
 # Server executable
 add_executable(pos_sp_server
